+++ example/staticgo/staticgo.go
@@ -3,7 +3,6 @@ package main
 import (
 	"flag"
 
-	"github.com/teambition/compressible-go"
 	"github.com/teambition/gear"
 	"github.com/teambition/gear/logging"
 	"github.com/teambition/gear/middleware/cors"
@@ -20,13 +19,13 @@ var (
 func main() {
 	flag.Parse()
 	app := gear.New()
-	app.Set(gear.SetCompress, compressible.WithThreshold(1024))
+	app.Set(gear.SetCompress, gear.ThresholdCompress(128))
 
 	app.UseHandler(logging.Default(true))
 	app.Use(cors.New())
 	app.Use(static.New(static.Options{Root: *path}))
 
-	logging.Println("staticgo v1.1.0, created by https://github.com/teambition/gear")
+	logging.Println("staticgo v1.2.0, created by https://github.com/teambition/gear")
 	logging.Printf("listen: %s, serve: %s\n", *address, *path)
 
 	if *certFile != "" && *keyFile != "" {